extracted cafeteria RPC logic into an individual file
